backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
groq==1.1.1
python-dotenv==1.0.1
